rcar_flash: initial commit
This is the initial commit for rcar_flash tool: universal automatic
tool for flashing Renesas Gen3/Gen4 boards via serial port.

There are lot to do. Right now it was tested with S4 and it is working
just fine. Features that are working:

 - Send flash writer image if requested
 - Flash loaders to QSPI
 - Flash loaders to eMMC
 - List known loaders for board
 - List know boards

What does not work:

 - CPLD communication. You need to switch serial download mode
   manually
 - Dynamic communication speed upgrade. There is room to improve
    flashing speed

Use `rcar_flash.py flash -h` to get help on flashing. Or
`rcar_flash.py -h` to get list of supported commands.

Signed-off-by: Volodymyr Babchuk <volodymyr_babchuk@epam.com>