Improve the readability of the progress
For 1MB+ files it's hard to estimate the progress
due to the long sequence of constantly changing digits.

This patch adds `_` as a separator for every three digits,
and this significantly improves visual understanding of
the progress.
Also, we print the percentage of the progress as suggested
by Volodymyr Babchuk <volodymyr_babchuk@epam.com>

So, instead of
1234567/12345678
we now see
1_234_567/12_345_678 (9%)

Signed-off-by: Ruslan Shymkevych <ruslan_shymkevych@epam.com>
Suggested-by: Volodymyr Babchuk <volodymyr_babchuk@epam.com>
Reviewed-by: Volodymyr Babchuk <volodymyr_babchuk@epam.com>