Rework handling of received chars
Received chars are printed immediately, so the user can
see long-running processes, like erase of HyperFlash.

Receive timeout now applies on a per-character basis,
not to the whole `read_until` sequence.

Also, print 'new line' before output of progress of sending
of a file, because sometimes the progress overwrites the
output from the board, due to manipulations with the cursor
position.

Signed-off-by: Ruslan Shymkevych <ruslan_shymkevych@epam.com>
Reviewed-by: Volodymyr Babchuk <volodymyr_babchuk@epam.com>