Add a `timeout` option for `wait_for`
In some cases, we need to wait extra time due to
the absence of any reaction from the flash_writer for
quite a long time. Eg.: flashing of a 16MB file results
in silence during 30 seconds.

In such cases, we need to temporarily change the read timeout.
This possibility is provided by the optional `timeout` keyword:

```
  - wait_for: "complete!"
    timeout: 40
    send: const
    val: "\r"
```

Signed-off-by: Ruslan Shymkevych <ruslan_shymkevych@epam.com>
Reviewed-by: Volodymyr Babchuk <volodymyr_babchuk@epam.com>